+++ .github/workflows/publish.yml
@@ -1,0 +1,30 @@
+name: Publish to PyPI
+
+on:
+  release:
+    types: [published]
+
+permissions:
+  contents: read
+  id-token: write # Required for uv publish to use PyPI OIDC
+
+jobs:
+  publish-common:
+    name: Build and Publish packages
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+        with:
+          fetch-tags: true
+          fetch-depth: 0
+
+      - name: Install uv
+        uses: astral-sh/setup-uv@v5
+        with:
+          enable-cache: true
+
+      - name: Build
+        run: uv build --package wren-pde-solver && uv build --package wren-ttd && uv build --package wren-commons
+
+      - name: Publish
+        run: uv publish
