--- packages/pde_solver/tests_solver/solver_tester.py
@@ -1,13 +1,0 @@
-from pde_solver.abc.solver import Solver
-
-
-class SolverTester:
-    """Class for testing solvers."""
-
-    def __init__(self, solver: Solver) -> None:
-        """Initialize the solver tester with given solver."""
-        self.solver: Solver = solver
-
-    """
-    TODO: actually implement this.
-    """
