--- packages/pde_ttd/src/pde_ttd/__init__.py
@@ -1,4 +1,0 @@
-from pde_ttd.core import TTD
-from pde_ttd.math import DEFAULT_EPSILON
-
-__all__ = ["DEFAULT_EPSILON", "TTD"]
