+++ packages/wren_common/pyproject.toml
@@ -1,0 +1,12 @@
+[project]
+name = "wren-common"
+version = "0.1.0"
+description = "Common code for the PDE library"
+dependencies = ["numpy~=2.3"]
+
+[build-system]
+requires = ["uv_build~=0.11.18"]
+build-backend = "uv_build"
+
+[tool.pytest.ini_options]
+pythonpath = ["src"]
