+++ packages/wren_pde_solver/pyproject.toml
@@ -1,0 +1,13 @@
+[project]
+name = "wren-pde-solver"
+version = "0.1.0"
+description = "Library providing solvers for PDEs"
+dependencies = ["numpy~=2.3", "sympy~=1.14", "wren_common"]
+
+
+[build-system]
+requires = ["uv_build~=0.11.18"]
+build-backend = "uv_build"
+
+[tool.pytest.ini_options]
+pythonpath = ["src"]
