+++ packages/wren_pde_solver/tests_solver/boundary_test.py
@@ -1,0 +1,61 @@
+from copy import deepcopy
+
+import numpy as np
+import pytest
+from wren_common.tests import (
+    SMALL_TEST_SCALARS,
+    TEST_TENSORS,
+    tensor_boundary,
+    tensor_interior,
+)
+from wren_pde_solver.boundary_conditions import ConstantDirichletBoundaryCondition
+from wren_pde_solver.pde_types import NDArray
+
+
+@pytest.mark.parametrize("value", SMALL_TEST_SCALARS)
+@pytest.mark.parametrize("tensor", deepcopy(TEST_TENSORS))
+def test_circular_add_and_remove_boundary(value: float, tensor: NDArray) -> None:
+    """Test adding and removing boundary to a tensor doesn't change anything."""
+    boundary = ConstantDirichletBoundaryCondition(value=value)
+    new_tensor = boundary.remove_boundary(boundary.apply_to_initial_condition(tensor))
+    np.testing.assert_array_equal(new_tensor, tensor)
+
+
+@pytest.mark.parametrize("value", SMALL_TEST_SCALARS)
+@pytest.mark.parametrize("tensor", deepcopy(TEST_TENSORS))
+def test_dirichlet_boundary_setup_initial_condition(
+    value: float, tensor: NDArray
+) -> None:
+    """
+    Test Dirichlet boundary condition for setup initial.
+
+    Should increase tensor size by 2 in each axis and adds the given value at boundary.
+    """
+    boundary = ConstantDirichletBoundaryCondition(value=value)
+    new_tensor = boundary.apply_to_initial_condition(tensor)
+    np.testing.assert_array_equal(new_tensor.shape, np.array(tensor.shape) + 2)
+    np.testing.assert_equal(tensor_boundary(new_tensor), value)
+    np.testing.assert_equal(
+        tensor_interior(new_tensor), tensor.reshape(-1)
+    )  # interior is returned as a 1D array
+
+
+@pytest.mark.parametrize("value", SMALL_TEST_SCALARS)
+@pytest.mark.parametrize("time", SMALL_TEST_SCALARS)
+@pytest.mark.parametrize("delta_time", SMALL_TEST_SCALARS)
+@pytest.mark.parametrize("tensor", deepcopy(TEST_TENSORS))
+def test_dirichlet_boundary_call(
+    value: float, time: float, delta_time: float, tensor: NDArray
+) -> None:
+    """
+    Test Dirichlet boundary condition for a time-step call.
+
+    Should keep tensor size and set zero (not the given value) at boundary.
+    """
+    boundary = ConstantDirichletBoundaryCondition(value=value)
+    new_tensor = boundary(tensor, time=time, delta_time=delta_time)
+    np.testing.assert_array_equal(new_tensor.shape, tensor.shape)
+    np.testing.assert_equal(tensor_boundary(new_tensor), 0)
+    np.testing.assert_equal(
+        tensor_interior(new_tensor), tensor_interior(tensor)
+    )  # interior is returned as a 1D array
